(CFACT-211) Query Ruby for correct library location
Without this change, lookup for libruby would search based on the
location of Ruby found on PATH. This required special handling for
wrappers like rbenv.

Change lookup to query Ruby's RbConfig for the location of its dynamic
library instead of searching the filesystem.
